tempest: make the ec2-api EBS image setup configurable
Setting up an EBS image for ec2-api in tempest can be
configured by means of a new tempest.ec2_api.enable_ebs_image
barclamp raw attribute. This attribute has a false default
value, which means that the EBS image creation needs to
be enabled explicitly.

The main reason for making this configurable and disabling it
by default is that the process of creating an EBS image
in the tempest cookbook recipe implies booting a nova
server, which has the unfortunate side effect of making the
tempest barclamp fail whenever there are problems with the
other barclamps, which is something to be avoided.