keystone: Use the v3 API for ec2 creds (SCRD-781)
The keystone v2.0 API is officially deprecated, and as of the Pike
release has started causing deprecation warnings in the logs. This patch
updates the add_ec2 command of the keystone_register helper to use the
v3 endpoint to avoid deprecation warnings.

It also rewrites the _build_ec2_obj method to write out a JSON-like
object to be consistent with the other _build_* methods, and updates
usage of the term "tenant" to "project" which is the v3 name for the
same thing. Unfortunately the body of the request still uses "tenant_id"
since this undocumented OS-EC2 keystone API is really just a mirror of
the v2.0 version of the same API.