keystone: Use v3 for service commands (SCRD-781)
The keystone v2.0 API is officially deprecated, and as of the Pike
release has started causing deprecation warnings in the logs. This patch
extracts requests to /services in the keystone_register provider into
its own method and converts it to use the v3 API endpoint, also taking
the opportunity to rewrite the _build_service_object method to write out
the body as a JSON-like object, which is more readable and consistent
with the _build_auth method. Also removes the private declaration from
the method, which is unneeded since it is declared earlier and applies
to every method below it.