keystone: Use v3 for project requests (SCRD-781)
The keystone v2.0 API is officially deprecated and as of the Pike
release has started causing deprecation warnings in the logs. This patch
extracts requests to /tenants in the keystone_register provider into its
own helper method and converts it to use the v3 API. It also renames the
add_tenant command to add_project since "project" is the v3 term for
what was known as a "tenant" in the v2.0 API and also renames the
tenant_name parameter for that command to project_name.

This patch also rewrites the _build_tenant_object method (now named
_build_project_object) to write out the body as a JSON-like object,
which is more readable and is consistent with the _build_auth method.
Additionally, the method can now accept a domain_id, but we will
continue just using the Default domain for now. Also removes the private
declaration from the method, which is unneeded since it is declared
above and applies to every method below it.