keystone: Use v3 for user requests (SCRD-781)
Keystone v2.0 is officially deprecated, and as of the Pike release has
started causing deprecation warnings in the logs. This patch updates
requests for /users in the keystone_register provider to use the v3 API.

Since an administrator can update a user's password with a generic
update request, and since changing the user's password via the change
password API in v3 requires knowing the original password, we can drop
the entire _build_user_password_object method and just send a PATCH
request with the regular user body.

This patch also updates the command's usage of the term "tenant" to
instead use "project" which is the v3 term for the same thing. We also
rewrite the _build_user_object method to write out the body as a
JSON-like object to improve readability and to be consistent with the
other _build_* methods.