keystone: Use v3 for role requests (SCRD-781)
The Keystone v2.0 API is officially deprecated and as of Pike has
started producing log warnings. This patch updates /roles requests in
the keystone_register command to use the v3 endpoint. Also taking the
opportunity to rewrite the _build_role_object method to make it
consistent with the other _build_* methods, and to remove the private
statement from it, which is unneeded since the private statement in ruby
applies to all methods below it. Additionally, remove unneeded
properties user_name and tenant_name in add_user calls.