keystone: Use v3 for role assignment (SCRD-781)
The keystone v2.0 API is officially deprecated, and as of the Pike
release has also started causing deprecation warnings in the logs. This
patch updates the add_access keystone_register command to use the v3
endpoint. Additionally it updates usage of the term "tenant" in the
add_access command to "project" which is the v3 name for the same thing,
and also renames some other variables to enhance readability. Also
remove the private declaration from the method since private is declared
earlier and applies to everything below it.