horizon: Reload horizon after migration instead of restart
Apache reload (graceful) will signal the apache process, that will
restart the child process (WSGI).  This can cause a race, were an
OpenStack client try to contact to one of the WSGI services, and
the old WSGI process can answer.  This process will fail because
the WSGI socket is not anymore in the file system.

The aproach here is to use a dummy service for 'horizon', what
will SIGUSR1 the wsgi:horizon childs from apache, signaling the
reload of the configuration file which was introduced on commit 0212068