keystone: Use v3 for endpoint creation (SRCD-781)
The keystone v2.0 API is officially deprecated, and as of the Pike
release has also started causing deprecation warnings in the logs. This
patch updates the add_endpoint_template command to use the v3 API. It
also renames it to just add_endpoint since an endpoint template is an
outdated concept from when endpoints needed to be specific to each
tenant.

This refactor borrows from the update_endpoint command which already
uses the v3 API for this. The difference is that add_endpoint needs to
handle both create and update, where update_endpoint is a special case
to handle updating the keystone endpoints.