Stop exposing passwords in the process table
Using --os-password in command line calls exposes the password in the
process table and may therefor potentially expose it to unprivileged
users. This patch converts all shell-outs to use the environment
variable version of their credential settings, which is not leaked to
the process table.

We can omit passing '--os-auth-type password' since it is always implied
by the use of '--os-password' or OS_PASSWORD.

For openstackclient commands, if we want to specify a particular
endpoint type, we should use --os-interface/OS_INTERFACE, not
--os-endpoint/OS_ENDPOINT.

Using OS_PROJECT_NAME instead of OS_TENANT_NAME is acceptable in all
cases except for the nova availability zone setter, which is a custom
script.

(cherry picked from commit a1accbafb6f46e6ab08862ce8b03ceea3f235423)