keystone: Rename tenant to project
In keystone v3 the name for what used to be called a 'tenant' is now
called a 'project'. Unfortunately we have been using the term 'tenant'
for a long time and it is deeply ingrained in the code base, but this
change initiates the switchover by renaming it in the keystone schema
and view so that end-users can see the proper name, and setting both
*_tenant and *_project in the keystone_settings helper so that future
keystone_settings calls can start using the new name.