Use keystone v3 in config auth urls (SCRD-781)
The keystone v2.0 API is officially deprecated, and as of Pike has
started issuing deprecation warnings in the logs when it is used. This
patch converts some services using the v2.0 identity endpoint for
authentication to use the v3 endpoint instead.

Unfortunately some services are not keystone-v3-ready, such as
glance-scrubber, trove, and the pacemaker compute resources, so just
leave them be for now.