horizon: add neutron FWaaS dashboard
Starting with Pike, the FWaaS horizon dashboard is
maintained as a separate project neutron-fwaas-dashboard [1]
that needs to be installed and configured separately [2].

Furthermore, the previously deprecated 'enable_firewall'
horizon configuration option is dropped in Pike. Horizon
has the ability to enable/disable the FWaaS panel automatically
based on the availability of the associated neutron service [1].

This commit also adds a policy_file.neutron_fwaas raw attribute
that controls the location of the FWaaS policy.json file used
by Horizon. It defaults to the files shipped with the
neutron-fwaas-dashboard package.

More information:
 - [1] Pike Horizon release notes: https://docs.openstack.org/releasenotes/horizon/pike.html
 - [2] neutron-fwaas-dashboard project: https://docs.openstack.org/neutron-fwaas-dashboard/pike/