keystone: Set an origin flag on apache2 restart
When we do a apache2 restart action, as the apache2 service is declared
on the apache2 cookbook, we have no idea of what triggered the restart
for it.

This makes it a problem to differentiate between the different cookbooks
calling the restart for the service, if we want to introduce a feature
to avoid restarts based on origin/barclamps.

This PR introduces a ruby_block that sets a temporal key on the node
when apache2 is restarted, to flag where the restart was triggered from.

The run_state values are ephemeral so we dont modify the node attributes
at all.

(cherry picked from commit 27150506dad11ff529da81ad261f5633c5a694ed)