keystone: Make password hash algorithm configurable
Since Pike, the new default password hash algorithm is
"bcrypt"[1]. This is more secure but also significantly slower (factor 10
in my tests) when eg. trying to get a new token via the v3 API.
To be able to increase performance (if wanted), make the hashing
algorithm and the hashing rounds for the selected algorithm
configurable. Reducing the value for "password_hash_rounds" increases
the performance.

[1]
https://docs.openstack.org/releasenotes/keystone/pike.html#new-features