rabbitmq: prevent template changes
As the rabbitmq backport has touched some configuration files, this
would mean that by only installing this update the user would get
service restarts. In order to prevent such behavior some extra
conditions are added to the templates - so if crowbar configuration is
not changed, template file contents will not differ.