manila, ha: fix bind_host in HA case
The bind_host IP address value reflected in the
osapi_share_listen configuration option (the listening
address for manila-api) was previously set to an empty
value in the HA case, which defaulted to the 0.0.0.0 value
being used by the manila-api service.
However, since Pike, the configured value is checked
against an IP address syntax and having an empty
value causes all manila services to fail.