mysql: Make sure galera resources are started on controller nodes only
Adding the location constraint should ensure that the service
is not running at remote cluster nodes.

mariadb: Set "clone-max" for the ms-galera resource

By setting this to the number of "real" cluster nodes (non-remote) we
avoid having the resource listed as "Stopped" on remote nodes
(compute-ha) in the crm status output. This is taken from the drbd
master/slave resource that we used for postgresql in the past.