keystone: Switch memcache backend to oslo_cache.memcache_pool
Per the oslo.cache documentation it's recommended to use the pooling
backend for eventlet based and threaded services. As our setup for the
keystone vhost in apache uses threads by default, let's use the pooling
backend. This seems to significantly reduce the number of open
connections to memcached.

Partial-Bug: bsc#1057822