database: Return hostname for listen address in case of SSL setup.
Hostname should be part of server certificate so we should make it
part of connection string so the connecting hostnames can match the
one in certificate.

But it has been seen that the vhostname might not be available at the
time of starting haproxy, which makes haproxy fail as it is not able
to resolve the address.

So let's keep it same as for other services and put only IP's into
haproxy config file.

(cherry picked from commit cdd4685a35e98f9039a9f0d2d13e838d1f7e7ef9)