Use httpchk in haproxy for galera
By using a script provided in the galera-python-clustercheck package we
are able to ensure a node is in sync with the galera cluster rather
than the previously superficial check of just ensuring mysql is running.

Install the package and configure it to use the monitoring user.

This requires a change to crowbar-ha in PR #256 to allow extending
the haproxy pacemaker group.