rabbitmq: Increase timeouts for start/promote actions (bsc#1059532)
When starting rabbitmq, especially if we are recovering from fencing,
the start could take a while as it needs to check for the tables
to be correct and/or reset the mnesia dir depending on the outcome, of
that check so the whole process could take much longer than the
default 60 seconds.

We also increase the promote timeout as it also needs to do some
internal checking for the master score and it could easily go over the
default timeout.

(cherry picked from commit 594853623315ba9be4c7563cda6e5a39dcf1293e)