Revert "neutron: Fix metadata server for ocata again"
The nova_metadata_ip option is deprecated in favor of the
new nova_metadata_host option because it reflects better
that the option accepts an IP address and also a DNS name. [1]

This change was originally included in b45377b, but then later
was reverted in 416c1bb  because it was breaking the (then
Ocata based) Devel:Cloud:8 build.
This PR reverts commit 416c1bb in effect bringing back these
suppressed changes.

More information:
[1] https://docs.openstack.org/releasenotes/neutron/pike.html#deprecation-notes