cinder,nova: use default key manager instead of barbican
Starting with Ocata, as part of the migration to castellan as
an external key manager interface lib, the default key manager
was changed from ConfKeyManager - the nova internal fixed-key,
insecure key manager - to Barbican [1]. This change isn't explicitly
documented in the release notes and creates a default dependency
between nova/cinder and Barbican. As currently nova/cinder have
not yet been fully integrated with Barbican, the encrypted volume
support is broken.

This is only a quick fix that brings back the ability to create
and manage encrypted volumes by switching back to the key
manager that was used before Ocata. It also fixes the following
tempest test cases:

  tempest.scenario.test_encrypted_cinder_volumes.TestEncryptedCinderVolumes.test_encrypted_cinder_volumes_cryptsetup
  tempest.scenario.test_encrypted_cinder_volumes.TestEncryptedCinderVolumes.test_encrypted_cinder_volumes_luks
  tempest.scenario.test_volume_boot_pattern.TestVolumeBootPattern.test_boot_server_from_encrypted_volume_luks

More info:
  - Castellan as a key manager lib: https://specs.openstack.org/openstack/nova-specs/specs/newton/implemented/use-castellan-key-manager.html