Revert "database: Fix HA service management"
This reverts commit b492d7dbcac413778f29a9d6e27e3c54c03240c0.

If crowbar restarts the database, it cannot guarantee it will do so
nondisruptively. When the database is deployed in an HA cluster, we must
assume that downtime for such a critical service is unacceptable. This
commit reverts back to the old behavior which requires a manual restart
of the database if any database parameters are changed by crowbar. This
should be the expected behavior even when we allow users to configure
mandatory service restarts[1].

[1] https://github.com/crowbar/crowbar-core/pull/1307