Revert "rabbitmq: Fix HA service management"
This reverts commit b08aaa1e90f7247d8111ec4203b68c59327eba71.

If crowbar restarts the message queue, it cannot guarantee it will do so
nondisruptively. When rabbitmq is deployed in an HA cluster, we must
assume that downtime for such a critical service is unacceptable. This
commit reverts back to the old behavior which requires a manual restart
of rabbitmq if any rabbitmq parameters are changed by crowbar. This
should be the expected behavior even when we allow users to configure
mandatory service restarts[1].

[1] https://github.com/crowbar/crowbar-core/pull/1307

(cherry picked from commit 2600ec22e483bd937663d1eaf194009ccebfa800)