neutron: Fixes for ACI integration - updates for Newton
This commit fixes minor issues encountered while testing ACI
integration with SUSE OpenStack Cloud 7. The ACI changes were
so far only being tested for Cloud 6 on the Cisco Lab.
Also recent changes in the crowbar framework caused failures while
applying the barclamp by enabling APIC as the ML2 backend.
The fixes include:
 - config setting for of_interface
 - modified path for linking ml2_conf_cisco_apic.ini
These changes were tested with Cloud7 on the Cisco ACI lab for L2 and L3
functionality and also shared with the Sales Engineering team.
So the commit needs to be merged with Cloud7 so PTFs can be
generated if required by customers wanting to upgrade from
Cloud6 to Cloud7.