swift: make default replicas count match storage node count
The current default swift replica count value is 3, which results
in a "Replica count of 3.0 requires more than 2 devices" failure
when the swift barclamp is deployed by mkcloud on setups with 3
nodes.

This commit extends that behavior by computing the default value
of the replicas count automatically based on the number of
swift-storage nodes using the formula:

   min(1, max(<no-swift-storage-nodes>, 3))

This fix is meant mainly to fix the DVR SOC8 Jenkins job, which
is currently broken.