neutron-l3-ha-service: don't capture SIGTERM
The service captures SIGTERM when migrating routers from dead agents.
This is a process that can take long time and can prevent Pacemaker to
properly shutdown the service leading to node fencing.

This change delegates the SIGTERM handling to neutron-ha-tool, which
does so in a per router basis instead of for the whole batch.

Reference: https://bugzilla.suse.com/show_bug.cgi?id=1059530