neutron-l3-ha-service: tweak timeouts
We found out that default timeouts for neutron-l3-ha-service were bigger
than Pacemaker default timeout (60s). This lead to a situation where we
tried to gracefully shutdown the service, but since we allowed that much
time, Pacemaker kicked in and fenced the node.

Since Pacemaker will call systemd service to stop the l3-service + tool,
we are reducing the time systemd takes to send a kill signal. Now
systemd will kill the l3-service + tool before Pacemaker tries to fence
the node.

Pacemaker timeout is configured to 60s. We want systemd to send the kill
signal at 55s. The systemd timeout is calculated as the maximum
l3-service :kill timeout + 5s. That's why we are changing the :kill
timeouts to 50s.