horizon: Explicit set REST_API_REQUIRED_SETTINGS (bsc#1046616)
SOC6 updated local_settings.py via Crowbar. Because of that, the file will
not be updated during a RPM package update (due to "config(noreplace)").
So explicitly set REST_API_REQUIRED_SETTINGS to get the correct settings.

Note: This fix is not needed for SOC8 because SOC7 will put the config
file into local_settings.d/ as a snippet. So local_settings.py will be
updated (if no user edited the file manually) during a SOC7->SOC8 update.