memcached: Make config non-HA-aware (bsc#1038223)
Without this patch, when deployed in an HA configuration, all of the
barclamps set their cache servers to all of the memcached servers in the
cluster in lexicographical order. This is not actually an optimal way to
configure memcached servers since if part of the cluster is down, the
memcached servers living on it will be inaccessible. python-memcached is
not tied to pacemaker and has no way of knowing that, so it passes the
entire list to python-memcached which attempts to connect to each server
serially, not attempting the next one until the first times out. The
effect is that any query to the OpenStack service will take a very long
time if the first memcached server in the list is down.

This patch fixes the issue by only using the local memcached server
instead of using all in the cluster. This is done for all barclamps
using memcached except for swift, since swift has its own way of doing
HA without pacemaker and also implements its own memcached client, so we
might as well leave it alone.