neutron: fix HA neutron-agents_before_ha timeout
Lately, running openstack CLI commands takes a long time due to
the fact that the CLI data model needs to be loaded from an increasing
number of python packages. This has a negative impact on cookbook
recipes that make extensive use of the openstack CLI, one of which
is the post_install neutron recipe, which evaluates the need to create
and then conditionally creates default objects (networks, routers).

Based on Jenkins HA job logs, an openstack CLI command can take up to
10-15 seconds to complete, which means the founder node is kept
busy for up to 100-150 seconds while evaluating the need to create
these default objects. This delay needs to be addressed in the
neutron cookbook, otherwise it may negatively impact other cookbooks
that are not prepared to handle that long of a delay.