keystone: Lazy evaluate auth hash
If the keystone admin password is updated through the keystone barclamp,
the change happens at compile time and might not propagate in the right
order to other recipes. This patch ensures the update stores the entire
credentials hash on the node object, and then uses the lazy syntax to
evaluate that value at converge time so that an update to the password
during the keystone proposal run will propagate to the other proposals
before they need it.