crowbar-pacemaker: clone restarts by default
When STONITH is enabled in a cluster, the default behavior for the
primitives is to fence the node on failure. This is a desirable behavior
for stateful services because it can lead to a corrupt state. On
stateless services (those that we can clone) fencing may be a too
intrusive action. We are setting the default action for clones to
restart instead of [implicit] fencing.