chef: add name to metadata and remove recommends
There is no reason to not have the name in all the cookbooks metadata
and it collides when using tools as some tools use the metadata name for
identification instead of just resorting to the directory name.

Plus it makes all cookbooks have the similar base metadata

Also removes 2 instances of recommends in the metadata as they are
useless. recommends and suggests do nothing on anything related to
chef-client or chef-server and they are deprecated due to this[0]

[0] https://chef.github.io/chef-rfc/rfc085-remove-unused-metadata.html