ceilometer: configurable default_api_return_limit (bsc#1046787)
The ceilometer admin dashboard in horizon doesn't specify
a limit for the number of entries returned by ceilometer API
calls, in which case the [api]/default_api_return_limit
configuration option is used to limit the number of returned
entries.
The default value of this option is 100, which may be too small
in cases where the ceilometer database has a large number of
entries (meters, samples, etc).

Making this option configurable as a raw attribute gives the
crowbar admin another parameter that can be used to adjust
ceilometer to larger environments.