neutron: fix additional external nets on compute nodes (bsc#1066171)
On compute nodes, the 'additional_external_networks' barclamp
attribute used to configure custom neutron external networks was
improperly taken from the local chef node instead of the remote
'neutron-network' node.

When used in combination with a DVR enabled setup, which requires
the external networks to be present on compute nodes too, this mishap
basically caused the custom external networks to be unreachable
without a clear error indication in the neutron logs.