neutron: Overwrite default config files for agents
In case a user already has a dhcp_agent.ini file with settings, or in
case the default dhcp_agent.ini file from the package contains
uncommented settings, then we need to invalidate these settings to
ensure that the config snippet we install for the agent will work.

Therefore we overwrite the file with some comment explaining what's
going on.