neutron: Fix config snippet support for metering agent
Because the metering agent is started with an explicit "--config-file
/etc/neutron/metadata_agent.ini" flag, options in this file will override
everything that is being set in config snippets.

So do not generate /etc/neutron/metadata_agent.ini (instead empty it), but
instead a config snippet.

This wasn't done as part of 654c0ea1 because the neutron package was missing a
/etc/neutron/neutron-metering-agent.conf.d directory back then.