nova: Fix apply with skip_unchanged_nodes
When there's absolutely no node impacted by the "apply" thanks to
skip_unchanged_nodes, then we hit some old micro-optimization in
apply_role_pre_chef_call that was not intended to be used that way: we
skip the whole method.

Unfortunately, we do need to run the method even with
skip_unchanged_nodes because we need to ensure that attributes, like the
HA related ones, are set.