nova: Do not reset sync marks uselessly with skip_unchanged_nodes
If skip_unchanged_nodes results in some control nodes to be skipped,
then we don't want to reset the sync marks as we will not run chef
through the apply. If we do the reset, then we will rely on periodic
chef runs to set the sync marks, which will take time to converge as
there's no guarantee that the periodic chef runs on the control nodes
are happening at the same time.

(cherry picked from commit e0194773b0daf234ed718e130c000ddfdf233f31)