rabbitmq: Set a different monitor interval for the Master
When using clustered RabbitMQ, we use a ms resource in pacemaker. As
explained in [1], we need to set a different monitor interval for slaves
and the master, otherwise one of them won't be monitored. Which is not
so great.

[1]https://clusterlabs.org/doc/en-US/Pacemaker/1.1/html/Pacemaker_Explained/_monitoring_multi_state_resources.html

(cherry picked from commit d8cc3390a9afd891d0e2ac06f97f9bdeca294558)