rabbitmq: allow to add extra users to rabbit
with the introduction of the definitions file it looks like
rabbit will disregard any existing users when restarting,
using the defintions file as the blank slate to start from.

This gets in the middle of having any extra users for other
tasks like monitoring and there is no mechanism to add users.

This PR adds an extra field to crowbar to add a list of users
that will get written into the definitions file, allowing
us to specify extra users to be created