galera: raise haproxy limit depending on max_connections (bsc#1068962)
Without this, haproxy is always limiting us to 1000 connections, not
exposing the full capabilities if db connections is set to 1000. Also
if db connections is set  < 1000, this allows the openstack services
to exceed the number of db connections, causing db healthcheck to fail
and pacemaker to go nuts. always reserve some spare connections
to avoid this from happening.