rabbitmq: Use pause_minority setting to deal with partitions better
We do care about data integrity, and the default setting is "ignore"
which can result in inconsistencies when the partition is over.

For two-node clusters, "pause_minority" doesn't work as it would require
the two nodes to always be up, which prevents any maintenance on a node.
But pacemaker should help here (with the fencing), so we keep the default
behavior and we don't add any safety net at the rabbitmq level.

See https://www.rabbitmq.com/partitions.html for more details