neutron: Fix for GBP authentication issue with ACI
With OpenStack Newton, the Group-Based-Policy library expects
special authentication mechanism with the Cisco ACI fabric.
It expects configuration of [apic_aim_auth] with similar
parameters as the [keystone_authtoken] section.
This commit adds this additional section to neutron.conf
along with the data from keystone_settings and a variable
in the recipe to add this section only when the APIC GBP
mode is used.