nova: set executor_thread_pool_size back to default
We raised this to a high value to workaround a scalability
issue in icehouse, but we don't need this anymore with newton+ it
seems. this higher value cloggs up a large amount of resources
without seeing a whole lot of performance benefit.

(cherry picked from commit 78732a60facfccad321f0de6c9899a6b5cb5d8b1)