neutron: Force IPv4 on Java based services
The Java based services midonet-cluster, cassandra, and zookeeper are
based on Java. Java does not necessarily create an IPv4 socket for the
servers and we might end up with only IPv6 sockets. This breaks the
OpenStack integration. This change depends on PR 1461 in crowbar-core
[1] which introduced a new Chef resource to override an environment
variable for a systemd service.

For a more detailed description see (expired) bug report for Ubuntu
[2].

[1] https://github.com/crowbar/crowbar-core/pull/1461
[2] https://bugs.launchpad.net/ubuntu/+source/zookeeper/+bug/888643