rabitmq: allow creating extra users
with the introduction of the definitions file it looks like
rabbit will disregard any existing users when reseting,
using the definitions file as the blank slate to start from.

This gets in the middle of having any extra users for other
tasks like monitoring and there is no mechanism to add users.

This patch allows to create an arbitrary number of users,
with permissions and tags so in case of a rabbit reset it
would keep those extra users intact.