rabbitmq: allow setting the log level
We may need to be able to set the log level for rabbitmq so
this patch allows to set it via the barclamp.

There is 4 different items in which the log level can be applied
(connection, channel, mirroring and federation). While on
rabbitmq 3.4 channel logging is not used, its ignored and
can be left there in case we upgrade to 3.5/3.6 later.